+++ packages/client/hmi-client/src/services/jupyter.ts
@@ -1,3 +1,4 @@
+import { EventEmitter, EventName, EventCallback } from '@/utils/emitter';
 import { SessionContext } from '@jupyterlab/apputils';
 import {
 	ServerConnection,
@@ -16,6 +17,7 @@ import * as kernel from '@jupyterlab/services/lib/kernel/kernel';
 import { KernelConnection as JupyterKernelConnection } from '@jupyterlab/services/lib/kernel';
 import API from '@/api/api';
 import { v4 as uuidv4 } from 'uuid';
+import { createMessage as createMessageWrapper } from '@jupyterlab/services/lib/kernel/messages';
 
 declare module '@jupyterlab/services/lib/kernel/messages' {
 	export function createMessage(options: JSONObject): JupyterMessage;
@@ -191,4 +193,110 @@ export const newSession = async (kernelName: string, name: string) => {
 export interface IMessageHistory {
 	message: string;
 	messageType: string;
+}
+
+class KernelMessage extends EventEmitter {
+	// Make fluent interface
+	register(eventName: EventName, fn: EventCallback): KernelMessage {
+		this.on(eventName, fn);
+		return this;
+	}
+}
+
+/**
+ * Provide a higher level interface for interacting with jupyter-kernel session
+ *
+ * Usage:
+ *   const manager = new KernelSessionManager(newSession('beaker', 'Beaker'));
+ *   await manager.init('beaker', 'Beaker', ...)
+ *   const msgHandle = manager.sendMessage( ... )
+ *   msgHandle.register('response_1', ...)
+ *   msgHandle.register('response_2', ...)
+ * */
+export class KernelSessionManager {
+	map: Map<string, KernelMessage>;
+
+	jupyterSession: SessionContext | null;
+
+	constructor() {
+		this.map = new Map();
+		this.jupyterSession = null;
+	}
+
+	async init(kernelName: string, name: string, context: any) {
+		const session = await newSession(kernelName, name);
+
+		// Dispatch
+		const iopubMessageHandler = (_session: any, message: any) => {
+			if (message.header.msg_type === 'status') {
+				return;
+			}
+			const msgType = message.header.msg_type;
+			const msgId = message.parent_header.msg_id;
+
+			if (this.map.has(msgId)) {
+				this.map.get(msgId)?.emit(msgType, message);
+			}
+		};
+
+		// Assign
+		this.jupyterSession = session;
+
+		// Setup context
+		session.kernelChanged.connect((_context, kernelInfo) => {
+			if (!kernelInfo.newValue) return;
+
+			const sessionKernel = kernelInfo.newValue;
+			if (sessionKernel.name === kernelName) {
+				session.iopubMessage.connect(iopubMessageHandler);
+
+				const messageBody = {
+					session: session.name || '',
+					channel: 'shell',
+					content: context,
+					msgType: 'context_setup_request',
+					msgId: createMessageId('context_setup')
+				};
+				sessionKernel.sendJupyterMessage(createMessageWrapper(messageBody));
+			}
+		});
+
+		// A bit of a hacky way to wait for the kernel to be setup in order
+		// to send custom messages
+		return new Promise((resolve) => {
+			const id = setInterval(() => {
+				console.log('waiting...', this.jupyterSession?.session?.kernel);
+				if (this.jupyterSession?.session?.kernel) {
+					clearInterval(id);
+					resolve(true);
+				}
+			}, 100);
+		});
+	}
+
+	sendMessage(msgType: string, messageContent: any) {
+		if (!this.jupyterSession) throw new Error('Session not ready');
+
+		const msgId = createMessageId(msgType);
+		const messageBody = {
+			session: this.jupyterSession.name || '',
+			channel: 'shell',
+			content: messageContent,
+			msgType,
+			msgId
+		};
+		const contextMessage = createMessageWrapper(messageBody);
+		const sessionKernel = this.jupyterSession.session?.kernel;
+		if (sessionKernel) {
+			const kernelMessage = new KernelMessage();
+			this.map.set(msgId, kernelMessage);
+			sessionKernel.sendJupyterMessage(contextMessage);
+			return kernelMessage;
+		}
+		return null;
+	}
+
+	disposeMessage(msgId: string) {
+		this.map.delete(msgId);
+	}
 }
